Reorganize package structure; rename some titles
