+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react'
 import ReactDOM from "react-dom"
 import CssBaseline from "@material-ui/core/CssBaseline";
 import { ThemeProvider } from "@material-ui/styles";
-import Dashboard from "./Dashboard";
+import Dashboard from "./view/budget-page/Budget";
 
 ReactDOM.render(
    <ThemeProvider>
