+++ src/view/budget-page/Budget.js
@@ -13,15 +13,17 @@ import Badge from '@material-ui/core/Badge';
 import Container from '@material-ui/core/Container';
 import Grid from '@material-ui/core/Grid';
 import Paper from '@material-ui/core/Paper';
-import Link from '@material-ui/core/Link';
 import MenuIcon from '@material-ui/icons/Menu';
 import ChevronLeftIcon from '@material-ui/icons/ChevronLeft';
 import NotificationsIcon from '@material-ui/icons/Notifications';
-import { mainListItems, secondaryListItems } from './listItems';
+import { mainListItems, secondaryListItems } from './../menu/listItems';
 import Chart from './Chart';
 import CurrentSpend from './CurrentSpend';
 import BudgetHistory from './BudgetHistory';
 import ForecastedSpend from './ForecastedSpend';
+
+// TODO: hit an endpoint to get the budget name
+var budgetName = "Monthly AWS Budget Alert io-example 012345678901"
 
 const drawerWidth = 240;
 
@@ -134,7 +136,7 @@ export default function Dashboard() {
             <MenuIcon />
           </IconButton>
           <Typography component="h1" variant="h6" color="inherit" noWrap className={classes.title}>
-            Dashboard
+            {budgetName}
           </Typography>
           <IconButton color="inherit">
             <Badge badgeContent={4} color="secondary">
